Chicken foot simulation data generator

chicken_generator generates chicken foot simulation data.

chicken_generator(seed = 1)

Arguments

seed
A scaler determining the seed to be used.

Value

A dataframe of the simulated data. The first column denotes the groups of the points and the last three columns are the coordinates of data points.

Details

chicken_generator generates chicken foot simulation data with 450 data points in total. It consists of three toes and one shank. Each of the toes contains 105 data points and the shank contains 135 data points.

Examples

chicken_generator(1)

#>     Group          X           Y            Z
#> 1   Toe 1 0.53070860 0.324379600 0.0695524374
#> 2   Toe 1 0.50825433 0.341739350 0.0981585418
#> 3   Toe 1 0.48801867 0.345069493 0.0621660686
#> 4   Toe 1 0.45612745 0.342912052 0.1480062050
#> 5   Toe 1 0.52296625 0.331216267 0.1033092990
#> 6   Toe 1 0.52324982 0.379195888 0.0900870461
#> 7   Toe 1 0.52125511 0.377379594 0.0911020295
#> 8   Toe 1 0.49493779 0.367987985 0.1250024302
#> 9   Toe 1 0.51480356 0.340593576 0.1206726557
#> 10  Toe 1 0.53273778 0.375225242 0.1126455491
#> 11  Toe 1 0.54500153 0.393218639 0.1396513549
#> 12  Toe 1 0.54322354 0.393493452 0.0880441047
#> 13  Toe 1 0.56224318 0.397939603 0.1053631772
#> 14  Toe 1 0.55943524 0.413327531 0.0897737786
#> 15  Toe 1 0.53450964 0.408856519 0.0430470896
#> 16  Toe 1 0.61185076 0.435739527 0.0666907100
#> 17  Toe 1 0.56486587 0.422332020 0.1045478368
#> 18  Toe 1 0.55131824 0.434983148 0.1188121305
#> 19  Toe 1 0.58361765 0.415465400 0.0877065760
#> 20  Toe 1 0.56951242 0.449476013 0.1227873655
#> 21  Toe 1 0.60565325 0.464689892 0.1764795302
#> 22  Toe 1 0.58192773 0.438226795 0.0902880984
#> 23  Toe 1 0.60835945 0.481636388 0.1160297766
#> 24  Toe 1 0.57744161 0.464218183 0.0926625498
#> 25  Toe 1 0.57104647 0.491021168 0.0654273174
#> 26  Toe 1 0.62672119 0.486364685 0.1019652722
#> 27  Toe 1 0.64133428 0.492118408 0.0760202415
#> 28  Toe 1 0.61396970 0.509549410 0.0917127712
#> 29  Toe 1 0.62746307 0.473690303 0.0942988758
#> 30  Toe 1 0.62540325 0.501796097 0.0708566087
#> 31  Toe 1 0.63156277 0.533054027 0.0725163023
#> 32  Toe 1 0.64482158 0.508696990 0.1343258991
#> 33  Toe 1 0.62574765 0.518032261 0.1326498109
#> 34  Toe 1 0.64714144 0.509386916 0.1163996441
#> 35  Toe 1 0.63087794 0.519357769 0.1477079127
#> 36  Toe 1 0.66003819 0.548931671 0.1540800876
#> 37  Toe 1 0.66149550 0.551242394 0.1169820732
#> 38  Toe 1 0.66231366 0.576380526 0.1238137603
#> 39  Toe 1 0.67779038 0.548733108 0.1052400708
#> 40  Toe 1 0.66435072 0.554617245 0.1510746461
#> 41  Toe 1 0.69302493 0.598557637 0.0738073977
#> 42  Toe 1 0.68003485 0.588121981 0.1390990505
#> 43  Toe 1 0.70245816 0.588598883 0.0840965231
#> 44  Toe 1 0.66143201 0.580050709 0.0853332491
#> 45  Toe 1 0.69795628 0.584794356 0.0881358093
#> 46  Toe 1 0.70071522 0.595417658 0.0374187095
#> 47  Toe 1 0.75751418 0.636443806 0.0810456871
#> 48  Toe 1 0.71609014 0.598727777 0.0730191489
#> 49  Toe 1 0.68999819 0.614674542 0.0686470050
#> 50  Toe 1 0.68290719 0.606015360 0.0896646033
#> 51  Toe 1 0.74817098 0.638284530 0.1075707207
#> 52  Toe 1 0.73488685 0.665828983 0.0910183554
#> 53  Toe 1 0.76084739 0.644309494 0.0804421266
#> 54  Toe 1 0.75740380 0.634180053 0.0588561553
#> 55  Toe 1 0.72726966 0.677366713 0.0391486244
#> 56  Toe 1 0.76288919 0.714918014 0.1695755051
#> 57  Toe 1 0.77934788 0.685688485 0.0826682390
#> 58  Toe 1 0.80511779 0.690349923 0.0413720643
#> 59  Toe 1 0.78414625 0.683260116 0.0865747085
#> 60  Toe 1 0.78775906 0.679850968 0.0990023213
#> 61  Toe 1 0.78396076 0.713905604 0.0481933158
#> 62  Toe 1 0.79555109 0.735452255 0.1264170041
#> 63  Toe 1 0.77286021 0.705844534 0.0994397114
#> 64  Toe 1 0.80412424 0.710302547 0.0398622887
#> 65  Toe 1 0.76332693 0.703720486 0.1100720649
#> 66  Toe 1 0.86141417 0.753175345 0.0546673726
#> 67  Toe 1 0.82477694 0.770157796 0.1425638189
#> 68  Toe 1 0.82226043 0.768800167 0.0558021906
#> 69  Toe 1 0.81115822 0.743349199 0.1062213670
#> 70  Toe 1 0.82163562 0.717409366 0.0782172031
#> 71  Toe 1 0.80065916 0.799641542 0.0858693090
#> 72  Toe 1 0.83539016 0.775372014 0.1065292847
#> 73  Toe 1 0.82273277 0.762834993 0.0775402486
#> 74  Toe 1 0.82852341 0.787994416 0.0432890357
#> 75  Toe 1 0.81220005 0.778876509 0.1265562284
#> 76  Toe 1 0.85730367 0.798113624 0.1553461846
#> 77  Toe 1 0.87049277 0.772853818 0.0799808233
#> 78  Toe 1 0.83552302 0.802697396 0.0349799399
#> 79  Toe 1 0.88121931 0.818770882 0.0986370227
#> 80  Toe 1 0.85916164 0.810405814 0.1009669827
#> 81  Toe 1 0.88591602 0.816378880 0.1166502536
#> 82  Toe 1 0.89133868 0.835696092 0.1281900932
#> 83  Toe 1 0.88179564 0.839533589 0.0995652670
#> 84  Toe 1 0.87823299 0.840722183 0.0606448654
#> 85  Toe 1 0.89711034 0.856471750 0.1327402780
#> 86  Toe 1 0.91251563 0.887278513 0.1321492903
#> 87  Toe 1 0.89866128 0.869489347 0.0068157409
#> 88  Toe 1 0.88331209 0.870881773 0.1118982860
#> 89  Toe 1 0.90048556 0.886036118 0.1049059863
#> 90  Toe 1 0.88830122 0.897480349 0.0766567093
#> 91  Toe 1 0.93900001 0.921385553 0.0825303263
#> 92  Toe 1 0.92287570 0.876000698 0.1056000341
#> 93  Toe 1 0.95753539 0.916219612 0.0615930095
#> 94  Toe 1 0.91699757 0.888913739 0.1852209708
#> 95  Toe 1 0.91941159 0.918592862 0.0971828146
#> 96  Toe 1 0.95764382 0.932934824 0.0563239347
#> 97  Toe 1 0.97027557 0.943108448 0.0680790538
#> 98  Toe 1 0.97406895 0.941353683 0.1250500570
#> 99  Toe 1 0.95394099 0.919182134 0.1069886050
#> 100 Toe 1 0.94883247 0.961718006 0.1271678648
#> 101 Toe 1 0.99976079 0.989913096 0.1232310946
#> 102 Toe 1 0.98085705 0.970942672 0.0546916059
#> 103 Toe 1 0.96978509 0.977370201 0.1612854939
#> 104 Toe 1 0.96574940 0.956242612 0.0781269438
#> 105 Toe 1 0.97815516 0.974128738 0.1500848249
#> 106 Toe 2 0.50958109 0.352629498 0.1294279084
#> 107 Toe 2 0.49337918 0.355139950 0.0745836436
#> 108 Toe 2 0.50386422 0.346083260 0.0874006369
#> 109 Toe 2 0.48315940 0.310451948 0.0459751402
#> 110 Toe 2 0.51404608 0.341667681 0.1091288862
#> 111 Toe 2 0.49133298 0.376724089 0.0801231665
#> 112 Toe 2 0.45524711 0.379562333 0.0362663974
#> 113 Toe 2 0.48145859 0.367056666 0.0786983387
#> 114 Toe 2 0.47073698 0.362321901 0.1398545885
#> 115 Toe 2 0.43120186 0.359967623 0.0798696144
#> 116 Toe 2 0.46454280 0.418289727 0.0419244912
#> 117 Toe 2 0.42912210 0.407716270 0.1240201621
#> 118 Toe 2 0.44949486 0.403770124 0.0849193076
#> 119 Toe 2 0.41420513 0.396167853 0.1200392308
#> 120 Toe 2 0.44870199 0.419670262 0.1583241311
#> 121 Toe 2 0.42090895 0.438581444 0.0001336105
#> 122 Toe 2 0.43425089 0.439459083 0.1380737096
#> 123 Toe 2 0.41221576 0.437993898 0.0692314805
#> 124 Toe 2 0.42731100 0.422856090 0.0764903506
#> 125 Toe 2 0.40531937 0.412360466 0.0856947797
#> 126 Toe 2 0.40363645 0.472305875 0.0913803317
#> 127 Toe 2 0.36215100 0.476936096 0.0962238738
#> 128 Toe 2 0.41166935 0.449930303 0.0865225872
#> 129 Toe 2 0.36664221 0.464365361 0.0858857247
#> 130 Toe 2 0.38853304 0.477512067 0.1539481881
#> 131 Toe 2 0.40235024 0.481644214 0.1384269051
#> 132 Toe 2 0.41878587 0.515055414 0.1134130729
#> 133 Toe 2 0.36877828 0.494510063 0.1302226918
#> 134 Toe 2 0.37302380 0.484420723 0.1025005527
#> 135 Toe 2 0.37251637 0.505396761 0.0877838191
#> 136 Toe 2 0.36740511 0.520489070 0.0784085454
#> 137 Toe 2 0.37729462 0.534302355 0.0787272878
#> 138 Toe 2 0.35189870 0.540570353 0.0812327936
#> 139 Toe 2 0.32830582 0.515392982 0.1425872202
#> 140 Toe 2 0.39113152 0.518897386 0.0670423063
#> 141 Toe 2 0.29231053 0.578770527 0.1423488350
#> 142 Toe 2 0.34303056 0.535638154 0.0665891532
#> 143 Toe 2 0.30234721 0.547022585 0.0893760766
#> 144 Toe 2 0.33250406 0.548859459 0.0295434782
#> 145 Toe 2 0.34487225 0.548329143 0.1117917646
#> 146 Toe 2 0.24888068 0.583340131 0.0846750736
#> 147 Toe 2 0.28174213 0.612626308 0.1231722892
#> 148 Toe 2 0.30059535 0.588856041 0.0907656900
#> 149 Toe 2 0.31382261 0.600417834 0.0688035898
#> 150 Toe 2 0.26219640 0.591821534 0.1144979464
#> 151 Toe 2 0.27609515 0.621574685 0.1256331165
#> 152 Toe 2 0.32849947 0.627587862 0.0703302917
#> 153 Toe 2 0.31009432 0.592981747 0.0650881790
#> 154 Toe 2 0.29708035 0.607496074 0.0367420702
#> 155 Toe 2 0.28525991 0.627938300 0.0403339253
#> 156 Toe 2 0.23485063 0.641761372 0.1015385113
#> 157 Toe 2 0.26853378 0.631068193 0.0667468218
#> 158 Toe 2 0.26591428 0.651100117 0.1138502455
#> 159 Toe 2 0.28513667 0.658407514 0.1439429711
#> 160 Toe 2 0.26068247 0.670487351 0.1299317551
#> 161 Toe 2 0.21549340 0.716757030 0.0790568221
#> 162 Toe 2 0.28304473 0.682071512 0.0511775029
#> 163 Toe 2 0.21959535 0.670431549 0.0734026162
#> 164 Toe 2 0.23954362 0.658097362 0.0788864449
#> 165 Toe 2 0.31357818 0.691115416 0.1746763811
#> 166 Toe 2 0.24530024 0.717038803 0.0525159717
#> 167 Toe 2 0.18592564 0.712072413 0.1025345144
#> 168 Toe 2 0.20840318 0.700307558 0.0525871439
#> 169 Toe 2 0.23660337 0.716162645 0.1039500652
#> 170 Toe 2 0.19109899 0.729174298 0.0608370711
#> 171 Toe 2 0.17472100 0.750813998 0.0023874652
#> 172 Toe 2 0.19529510 0.742135188 0.0586252007
#> 173 Toe 2 0.17176852 0.752707027 0.0575186339
#> 174 Toe 2 0.19364075 0.742883343 0.0933506536
#> 175 Toe 2 0.18852248 0.763537587 0.0774170082
#> 176 Toe 2 0.15015416 0.763644293 0.0321900409
#> 177 Toe 2 0.17162806 0.794989130 0.0865808191
#> 178 Toe 2 0.19120912 0.780521180 0.1333493776
#> 179 Toe 2 0.17760943 0.773473532 0.1157352948
#> 180 Toe 2 0.15941196 0.801264609 0.0967320914
#> 181 Toe 2 0.15158164 0.810324733 0.0768985826
#> 182 Toe 2 0.13713494 0.801903478 0.1042142650
#> 183 Toe 2 0.11927723 0.818330464 0.1098057553
#> 184 Toe 2 0.13373019 0.781974220 0.0923308026
#> 185 Toe 2 0.17265359 0.816401198 0.1382307350
#> 186 Toe 2 0.13966241 0.826795559 0.0654187977
#> 187 Toe 2 0.11276083 0.842627690 0.0093364522
#> 188 Toe 2 0.11451420 0.852234675 0.1203557044
#> 189 Toe 2 0.10519022 0.835129450 0.0739714854
#> 190 Toe 2 0.13573559 0.834873999 0.0736333415
#> 191 Toe 2 0.06624392 0.887753117 0.0565668297
#> 192 Toe 2 0.09714926 0.882295920 0.1216624209
#> 193 Toe 2 0.12550784 0.901531471 0.1201246530
#> 194 Toe 2 0.12737432 0.848609345 0.0751618412
#> 195 Toe 2 0.10356603 0.859193680 0.1216331095
#> 196 Toe 2 0.08524843 0.902491379 0.1227013999
#> 197 Toe 2 0.05353774 0.904594538 0.0865579257
#> 198 Toe 2 0.06114407 0.924884833 0.1082340435
#> 199 Toe 2 0.08797885 0.892620316 0.1049593451
#> 200 Toe 2 0.05459996 0.878320674 0.0647999278
#> 201 Toe 2 0.04924087 0.923024985 0.0796300012
#> 202 Toe 2 0.05260821 0.918248114 0.1447272697
#> 203 Toe 2 0.03824165 0.935147875 0.1103293113
#> 204 Toe 2 0.01795671 0.964681738 0.1108195697
#> 205 Toe 2 0.09950296 0.941055556 0.0868475373
#> 206 Toe 2 0.00000000 0.966450071 0.0767125598
#> 207 Toe 2 0.03633010 0.946123399 0.0811295394
#> 208 Toe 2 0.01709722 0.993529332 0.0283826690
#> 209 Toe 2 0.01562989 0.993249584 0.0829160867
#> 210 Toe 2 0.04930177 0.978994544 0.0829409571
#> 211 Shank 0.45176563 0.325934337 0.0483148183
#> 212 Shank 0.49363766 0.344772275 0.0736215944
#> 213 Shank 0.51472260 0.341444501 0.0658538225
#> 214 Shank 0.48726221 0.322343068 0.0923712209
#> 215 Shank 0.49972352 0.342692202 0.0361328572
#> 216 Shank 0.44161594 0.320214008 0.0354117030
#> 217 Shank 0.50050042 0.308475178 0.1216109506
#> 218 Shank 0.51065243 0.343691450 0.0263329123
#> 219 Shank 0.47874267 0.300426092 0.0802450062
#> 220 Shank 0.51455848 0.322778567 0.0826604322
#> 221 Shank 0.47728191 0.299260950 0.1033611574
#> 222 Shank 0.52410217 0.285771175 0.1261497599
#> 223 Shank 0.48989741 0.312675014 0.0618937688
#> 224 Shank 0.47129937 0.304487215 0.0574006762
#> 225 Shank 0.51191934 0.332605734 0.0291175239
#> 226 Shank 0.49187368 0.301983141 0.0410583864
#> 227 Shank 0.48731941 0.271762799 0.0544131052
#> 228 Shank 0.51965323 0.290552239 0.0657861338
#> 229 Shank 0.53159464 0.289576008 0.0991212846
#> 230 Shank 0.47962175 0.273531088 0.0563999067
#> 231 Shank 0.48629092 0.273738520 0.1245340596
#> 232 Shank 0.53991078 0.239395667 0.0890095967
#> 233 Shank 0.49865085 0.279515099 0.0222082478
#> 234 Shank 0.50017153 0.280507035 0.0649813123
#> 235 Shank 0.47796740 0.267770764 0.1079578123
#> 236 Shank 0.45216970 0.276765160 0.0442098919
#> 237 Shank 0.49417187 0.273160086 0.0378881483
#> 238 Shank 0.53367260 0.252794602 0.0767732214
#> 239 Shank 0.50307294 0.239700052 0.1397306272
#> 240 Shank 0.52238500 0.251597158 0.1536036439
#> 241 Shank 0.49067674 0.241021797 0.1089533194
#> 242 Shank 0.51919090 0.219927525 0.0641689026
#> 243 Shank 0.47918921 0.251548885 0.1532422088
#> 244 Shank 0.48076205 0.246741256 0.0889057317
#> 245 Shank 0.50894290 0.235374908 0.0572250386
#> 246 Shank 0.47351759 0.194419677 0.0757369829
#> 247 Shank 0.52113967 0.206495222 0.1243402814
#> 248 Shank 0.48420532 0.241245936 0.0846600523
#> 249 Shank 0.53651150 0.233388582 0.1338662509
#> 250 Shank 0.53590741 0.211383791 0.1091723858
#> 251 Shank 0.50022629 0.181344695 0.0403057594
#> 252 Shank 0.51285419 0.235184720 0.0618120417
#> 253 Shank 0.47480958 0.221707646 0.0475168718
#> 254 Shank 0.49796513 0.219201038 0.1152369778
#> 255 Shank 0.50396933 0.199542209 0.0460941526
#> 256 Shank 0.48082046 0.187982506 0.1514488720
#> 257 Shank 0.52221011 0.170394589 0.1143486935
#> 258 Shank 0.50324054 0.204819714 0.1369518797
#> 259 Shank 0.49272129 0.211296617 0.0966367259
#> 260 Shank 0.52293932 0.191658551 0.0523891070
#> 261 Shank 0.49871114 0.158903834 0.0648421121
#> 262 Shank 0.52811687 0.181911052 0.1322457183
#> 263 Shank 0.50382420 0.175145657 0.0619242649
#> 264 Shank 0.51692727 0.185595914 0.1054953050
#> 265 Shank 0.49603112 0.166159674 0.1231500021
#> 266 Shank 0.49689065 0.146335893 0.0553239190
#> 267 Shank 0.49192655 0.136524818 0.0912879317
#> 268 Shank 0.52039903 0.173972853 0.0484169204
#> 269 Shank 0.51501535 0.160605015 0.0000000000
#> 270 Shank 0.46589558 0.152803576 0.1329577302
#> 271 Shank 0.49696855 0.171173877 0.1041529269
#> 272 Shank 0.51969118 0.155474033 0.1173389424
#> 273 Shank 0.50633267 0.122210904 0.1152713331
#> 274 Shank 0.50207318 0.122816253 0.0060663828
#> 275 Shank 0.50412785 0.151416670 0.0833628110
#> 276 Shank 0.52377743 0.121809106 0.0975445276
#> 277 Shank 0.54520946 0.152926796 0.0468417284
#> 278 Shank 0.50650155 0.142458734 0.1746389149
#> 279 Shank 0.47958752 0.121551646 0.0418130837
#> 280 Shank 0.48076762 0.131760422 0.0910747958
#> 281 Shank 0.50946679 0.115260878 0.1624995607
#> 282 Shank 0.51365221 0.125558462 0.0185578355
#> 283 Shank 0.47980939 0.117189020 0.1995779415
#> 284 Shank 0.45299882 0.087205551 0.0824449417
#> 285 Shank 0.49395606 0.117862074 0.0756282300
#> 286 Shank 0.52844724 0.107871834 0.0583287148
#> 287 Shank 0.49932989 0.110398163 0.0646286746
#> 288 Shank 0.46645111 0.086759444 0.0579939394
#> 289 Shank 0.50509233 0.093206332 0.0776194284
#> 290 Shank 0.49806861 0.102102743 0.0806840597
#> 291 Shank 0.48162684 0.083910233 0.1542639069
#> 292 Shank 0.51315491 0.083782582 0.1038006433
#> 293 Shank 0.49920144 0.093597057 0.0696726403
#> 294 Shank 0.50411136 0.085690697 0.0761533526
#> 295 Shank 0.52254015 0.076573937 0.0738307274
#> 296 Shank 0.50018677 0.068528219 0.1535415222
#> 297 Shank 0.48180958 0.061901278 0.0579926260
#> 298 Shank 0.52476265 0.084082640 0.1497294156
#> 299 Shank 0.45865638 0.056830938 0.0610676936
#> 300 Shank 0.48992175 0.074343936 0.1010791527
#> 301 Shank 0.50236047 0.036002690 0.0372438185
#> 302 Shank 0.49118928 0.058461992 0.0985355938
#> 303 Shank 0.47105411 0.052869111 0.1010110133
#> 304 Shank 0.50350791 0.059206269 0.0521690259
#> 305 Shank 0.53595185 0.037255891 0.1146601533
#> 306 Shank 0.48194816 0.043633155 0.0687890948
#> 307 Shank 0.48029690 0.064659216 0.0846177957
#> 308 Shank 0.49665996 0.038479183 0.1160122750
#> 309 Shank 0.51914332 0.018797603 0.0929501073
#> 310 Shank 0.50733503 0.039440494 0.1603037839
#> 311 Shank 0.49961871 0.046774528 0.1204183773
#> 312 Shank 0.51398366 0.018758733 0.1239962090
#> 313 Shank 0.51779278 0.000000000 0.0856053064
#> 314 Shank 0.50515882 0.013105299 0.0068462893
#> 315 Shank 0.48175975 0.005579936 0.1197787570
#> 316 Toe 3 0.50382458 0.351113588 0.1203912539
#> 317 Toe 3 0.51976007 0.333724774 0.1392854880
#> 318 Toe 3 0.50863330 0.342236152 0.1711920969
#> 319 Toe 3 0.49111686 0.338115715 0.0742419122
#> 320 Toe 3 0.45562544 0.333485142 0.1084148703
#> 321 Toe 3 0.50957591 0.351591598 0.0635985127
#> 322 Toe 3 0.51802642 0.386346872 0.0909690851
#> 323 Toe 3 0.48126954 0.337932013 0.0028845008
#> 324 Toe 3 0.48632541 0.376807877 0.1697467527
#> 325 Toe 3 0.44047164 0.362432247 0.0391026511
#> 326 Toe 3 0.50304227 0.404881022 0.0577371136
#> 327 Toe 3 0.48413158 0.399719405 0.1050989253
#> 328 Toe 3 0.49964851 0.392265083 0.1095086574
#> 329 Toe 3 0.51609122 0.408667180 0.1632530266
#> 330 Toe 3 0.51399048 0.404888666 0.0939558451
#> 331 Toe 3 0.52316707 0.426297117 0.0746303403
#> 332 Toe 3 0.52270609 0.418766299 0.0680460420
#> 333 Toe 3 0.48302448 0.424890256 0.1097817505
#> 334 Toe 3 0.50466962 0.410774866 0.1287127875
#> 335 Toe 3 0.50184094 0.421199940 0.0959485382
#> 336 Toe 3 0.49695872 0.455005033 0.0329623765
#> 337 Toe 3 0.48861532 0.459312088 0.1203060748
#> 338 Toe 3 0.47375308 0.440656041 0.0241450921
#> 339 Toe 3 0.47849093 0.442925373 0.0476484523
#> 340 Toe 3 0.54120230 0.466841680 0.1269205127
#> 341 Toe 3 0.48977390 0.477290090 0.0801447570
#> 342 Toe 3 0.49247904 0.480513472 0.1190053676
#> 343 Toe 3 0.48464976 0.503885868 0.1193938409
#> 344 Toe 3 0.49584607 0.476546833 0.1186254417
#> 345 Toe 3 0.51800790 0.500914687 0.1020840094
#> 346 Toe 3 0.51808981 0.531886042 0.1076959395
#> 347 Toe 3 0.51530397 0.529334027 0.0834649782
#> 348 Toe 3 0.50340558 0.512661499 0.1223953252
#> 349 Toe 3 0.51016656 0.513342628 0.1501504497
#> 350 Toe 3 0.50349129 0.551122379 0.0914488399
#> 351 Toe 3 0.45399015 0.554035525 0.0931567968
#> 352 Toe 3 0.52775332 0.573468520 0.0995632152
#> 353 Toe 3 0.52311231 0.549044204 0.0559922529
#> 354 Toe 3 0.48495504 0.580082057 0.1764655587
#> 355 Toe 3 0.47222507 0.552135285 0.1199925687
#> 356 Toe 3 0.48126794 0.600453727 0.0268930045
#> 357 Toe 3 0.46834368 0.578215015 0.0821743482
#> 358 Toe 3 0.46351436 0.587457820 0.0269141565
#> 359 Toe 3 0.51157450 0.592482336 0.0629132903
#> 360 Toe 3 0.49924886 0.611913748 0.0974497887
#> 361 Toe 3 0.53610720 0.612558395 0.1189376072
#> 362 Toe 3 0.51468619 0.637577668 0.0867004266
#> 363 Toe 3 0.49369358 0.615441144 0.1055546497
#> 364 Toe 3 0.50026897 0.629676961 0.0995913328
#> 365 Toe 3 0.49794415 0.623012250 0.0925050524
#> 366 Toe 3 0.50855509 0.655766269 0.0178418995
#> 367 Toe 3 0.48055583 0.637520440 0.1516120167
#> 368 Toe 3 0.47872847 0.672331152 0.1292414505
#> 369 Toe 3 0.49948055 0.649260591 0.0646030156
#> 370 Toe 3 0.51198175 0.654992290 0.1051631685
#> 371 Toe 3 0.49845996 0.693270719 0.0942808387
#> 372 Toe 3 0.49164884 0.673801130 0.1166020494
#> 373 Toe 3 0.48607191 0.678102539 0.1035025377
#> 374 Toe 3 0.48935245 0.682892465 0.1109324426
#> 375 Toe 3 0.52537908 0.662412206 0.0421344048
#> 376 Toe 3 0.52354987 0.736484079 0.0472187793
#> 377 Toe 3 0.47646482 0.703243871 0.0840680944
#> 378 Toe 3 0.49193170 0.704436300 0.0043839664
#> 379 Toe 3 0.52779714 0.697461990 0.0678675101
#> 380 Toe 3 0.48675666 0.718399315 0.0739719611
#> 381 Toe 3 0.46294205 0.741261423 0.1158805076
#> 382 Toe 3 0.49759200 0.735742621 0.1054278663
#> 383 Toe 3 0.51590083 0.742559314 0.0455776718
#> 384 Toe 3 0.47725368 0.758546426 0.1346444200
#> 385 Toe 3 0.49569047 0.740252053 0.1197528793
#> 386 Toe 3 0.47492206 0.770054538 0.0485157172
#> 387 Toe 3 0.51572075 0.791036647 0.0957983935
#> 388 Toe 3 0.49231014 0.780865433 0.0869949197
#> 389 Toe 3 0.47626901 0.776988738 0.0904842642
#> 390 Toe 3 0.49165954 0.736821216 0.0696981837
#> 391 Toe 3 0.54419105 0.827738333 0.0784112530
#> 392 Toe 3 0.50019533 0.806820187 0.0739668335
#> 393 Toe 3 0.49434167 0.822405528 0.1092100944
#> 394 Toe 3 0.48875972 0.798176684 0.1453790012
#> 395 Toe 3 0.51588811 0.809350634 0.1265651821
#> 396 Toe 3 0.49428281 0.857945966 0.1660473792
#> 397 Toe 3 0.50076259 0.837334589 0.1062278052
#> 398 Toe 3 0.49159960 0.829365437 0.0824479685
#> 399 Toe 3 0.46766926 0.859097704 0.1649030704
#> 400 Toe 3 0.48761518 0.819038850 0.0641717301
#> 401 Toe 3 0.46861897 0.886700228 0.0366917806
#> 402 Toe 3 0.52656005 0.872054344 0.1595843857
#> 403 Toe 3 0.50437502 0.871910567 0.0261054759
#> 404 Toe 3 0.49366051 0.884279213 0.0173026036
#> 405 Toe 3 0.52577245 0.893220926 0.1163090646
#> 406 Toe 3 0.50825215 0.899519488 0.1248589746
#> 407 Toe 3 0.48632688 0.929708079 0.1368007253
#> 408 Toe 3 0.49938777 0.908530892 0.0813923085
#> 409 Toe 3 0.55261701 0.905658462 0.1190818970
#> 410 Toe 3 0.49899921 0.913405340 0.1021163860
#> 411 Toe 3 0.50183492 0.949365763 0.0812118028
#> 412 Toe 3 0.54829667 0.939690309 0.1176043706
#> 413 Toe 3 0.47119642 0.944513755 0.1058508046
#> 414 Toe 3 0.48074233 0.934246117 0.0726652818
#> 415 Toe 3 0.47371252 0.916184383 0.0534406922
#> 416 Toe 3 0.49038251 0.970707748 0.0787802322
#> 417 Toe 3 0.51260958 0.983583375 0.1404555737
#> 418 Toe 3 0.50025110 0.999841357 0.0862598088
#> 419 Toe 3 0.50596249 0.987291457 0.0826490080
#> 420 Toe 3 0.47418795 0.958293628 0.0461698114
#> 421 Shank 0.45963981 0.377336348 0.0768728162
#> 422 Shank 0.41180963 0.326122626 0.1226833547
#> 423 Shank 0.41160356 0.349670840 0.0892083065
#> 424 Shank 0.43166449 0.380538400 0.0348725678
#> 425 Shank 0.46076308 0.490648856 0.1063151693
#> 426 Shank 0.49990974 0.324914396 0.0647293313
#> 427 Shank 0.56106326 0.297357108 0.1391296689
#> 428 Shank 0.45173141 0.347341396 0.0567385669
#> 429 Shank 0.41844971 0.309127762 0.0162189672
#> 430 Shank 0.65509411 0.355843337 0.0556295250
#> 431 Shank 0.59554959 0.342649704 0.0675310366
#> 432 Shank 0.56712935 0.386395284 0.0568146366
#> 433 Shank 0.66311276 0.284271416 0.0816463016
#> 434 Shank 0.40689741 0.273242181 0.0665474811
#> 435 Shank 0.69649197 0.397676118 0.0743380176
#> 436 Shank 0.61780415 0.337446346 0.1055168950
#> 437 Shank 0.52982939 0.285948058 0.0425610502
#> 438 Shank 0.56019100 0.279156304 0.1237576801
#> 439 Shank 0.50632149 0.302552678 0.0915658679
#> 440 Shank 0.48984777 0.353832140 0.0900603037
#> 441 Shank 0.34963110 0.303678171 0.0057326175
#> 442 Shank 0.57150634 0.387547526 0.0789601112
#> 443 Shank 0.70543902 0.338279099 0.0682222256
#> 444 Shank 0.70041076 0.336126914 0.1274345584
#> 445 Shank 0.51669874 0.326641161 0.1503744945
#> 446 Shank 0.49012985 0.313984954 0.1436395265
#> 447 Shank 0.51786611 0.319813616 0.0446843904
#> 448 Shank 0.61009953 0.379229897 0.0580756117
#> 449 Shank 0.64180653 0.267265866 0.0878954770
#> 450 Shank 0.53948597 0.376101949 0.0469500244